Create Plant_Irrigation_Lite.ino
This is a more light version of the original code that simplifies certain processes. It uses the WiFi to find time and the system will work everyday at 17:00 regardless of the month. It will check the soil humidity and it will start watering the plants for 5 minutes. It will then check again the status of the soil humidity and if it is less than 80% it will continue to water them for another 5 minutes. The system will stop working after 11 minutes. It will then report the values found to the thingspeak channel and it will go into deep sleep waking up every half an hour to check the time.